Update for sample design change
