+++ src/Android/Xamarin.Android/ArcGISRuntime.Xamarin.Samples.Android.csproj
@@ -123,7 +123,7 @@
 	  <None Include="Samples\Map\MobileMapSearchAndRoute\MobileMapSearchAndRoute.jpg" />
 	  <None Include="Samples\Search\OfflineGeocode\OfflineGeocode.jpg" />
 	  <None Include="Samples\Search\ReverseGeocode\ReverseGeocode.jpg" />
-    <None Include="Samples\Layers\DisplayKmlTours\DisplayKmlTours.jpg" />
+    <None Include="Samples\Layers\PlayKmlTours\PlayKmlTours.jpg" />
     <None Include="Samples\Map\MapReferenceScale\MapReferenceScale.jpg" />
     <None Include="Samples\Data\EditFeatureAttachments\EditFeatureAttachments.jpg" />
     <None Include="Samples\Data\DeleteFeatures\DeleteFeatures.jpg" />
@@ -284,7 +284,7 @@
 	  <Compile Include="Samples\Layers\OpenStreetMapLayer\OpenStreetMapLayer.cs" />
 	  <Compile Include="Samples\Map\ManageOperationalLayers\ManageOperationalLayers.cs" />
 	  <Compile Include="Samples\Data\UpdateAttributes\UpdateAttributes.cs" />
-    <Compile Include="Samples\Layers\DisplayKmlTours\DisplayKmlTours.cs" />
+    <Compile Include="Samples\Layers\PlayKmlTours\PlayKmlTours.cs" />
     <Compile Include="Samples\Map\MapReferenceScale\MapReferenceScale.cs" />
     <Compile Include="Samples\Data\UpdateGeometries\UpdateGeometries.cs" />
     <Compile Include="Samples\Data\AddFeatures\AddFeatures.cs" />
