+++ src/Forms/Shared/Samples/Layers/PlayKmlTours/PlayKmlTours.xaml
@@ -2,7 +2,7 @@
              xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"             
              xmlns:esriUI="clr-namespace:Esri.ArcGISRuntime.Xamarin.Forms;assembly=Esri.ArcGISRuntime.Xamarin.Forms"
              xmlns:resources="clr-namespace:Forms.Resources;assembly=ArcGISRuntime"
-             x:Class="ArcGISRuntimeXamarin.Samples.DisplayKmlTours.DisplayKmlTours">
+             x:Class="ArcGISRuntimeXamarin.Samples.PlayKmlTours.PlayKmlTours">
     <RelativeLayout>
         <esriUI:SceneView x:Name="MySceneView" Style="{StaticResource SceneWithFormStyle}"
                         BindingContext="{x:Reference Name=ResponsiveFormContainer}" />
