+++ src/UWP/ArcGISRuntime.UWP.Viewer/ArcGISRuntime.UWP.Viewer.csproj
@@ -118,7 +118,7 @@
 	  <Content Include="Samples\Data\ViewPointCloudDataOffline\ViewPointCloudDataOffline.jpg" />
 	  <Content Include="Samples\Layers\AddAnIntegratedMeshLayer\AddAnIntegratedMeshLayer.jpg" />
 	  <Content Include="Samples\GraphicsOverlay\DictionaryRendererGraphicsOverlay\DictionaryRendererGraphicsOverlay.jpg" />
-    <Content Include="Samples\Layers\DisplayKmlTours\DisplayKmlTours.jpg" />
+    <Content Include="Samples\Layers\PlayKmlTours\PlayKmlTours.jpg" />
     <Content Include="Samples\Map\OpenMobileScenePackage\OpenMobileScenePackage.jpg" />
     <Content Include="Samples\Map\MapReferenceScale\MapReferenceScale.jpg" />
     <Content Include="Samples\Map\MobileMapSearchAndRoute\MobileMapSearchAndRoute.jpg" />
@@ -316,8 +316,8 @@
   </ItemGroup>
   <ItemGroup>
     <!-- Sample Code -->
-    <Compile Include="Samples\Layers\DisplayKmlTours\DisplayKmlTours.xaml.cs">
-      <DependentUpon>DisplayKmlTours.xaml</DependentUpon>
+    <Compile Include="Samples\Layers\PlayKmlTours\PlayKmlTours.xaml.cs">
+      <DependentUpon>PlayKmlTours.xaml</DependentUpon>
     </Compile>
     <Compile Include="Samples\Map\OpenMobileScenePackage\OpenMobileScenePackage.xaml.cs">
       <DependentUpon>OpenMobileScenePackage.xaml</DependentUpon>
@@ -792,7 +792,7 @@
   </ItemGroup>
   <ItemGroup>
     <!-- Sample XAML -->
-    <Page Include="Samples\Layers\DisplayKmlTours\DisplayKmlTours.xaml">
+    <Page Include="Samples\Layers\PlayKmlTours\PlayKmlTours.xaml">
       <Generator>MSBuild:Compile</Generator>
       <SubType>Designer</SubType>
     </Page>
