+++ src/UWP/ArcGISRuntime.UWP.Viewer/Samples/Layers/PlayKmlTours/PlayKmlTours.jpg
