+++ src/UWP/ArcGISRuntime.UWP.Viewer/Samples/Layers/PlayKmlTours/PlayKmlTours.xaml
@@ -1,5 +1,5 @@
 ﻿<UserControl
-    x:Class="ArcGISRuntime.UWP.Samples.DisplayKmlTours.DisplayKmlTours"
+    x:Class="ArcGISRuntime.UWP.Samples.PlayKmlTours.PlayKmlTours"
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
     xmlns:esri="using:Esri.ArcGISRuntime"
