+++ src/WPF/ArcGISRuntime.WPF.Viewer/ArcGISRuntime.WPF.Viewer.csproj
@@ -94,8 +94,8 @@
   </ItemGroup>
   <ItemGroup>
     <!-- Sample Code -->
-    <Compile Include="Samples\Layers\DisplayKmlTours\DisplayKmlTours.xaml.cs">
-      <DependentUpon>DisplayKmlTours.xaml</DependentUpon>
+    <Compile Include="Samples\Layers\PlayKmlTours\PlayKmlTours.xaml.cs">
+      <DependentUpon>PlayKmlTours.xaml</DependentUpon>
     </Compile>
     <Compile Include="Samples\Map\OpenMobileScenePackage\OpenMobileScenePackage.xaml.cs">
       <DependentUpon>OpenMobileScenePackage.xaml</DependentUpon>
@@ -590,7 +590,7 @@
   </ItemGroup>
   <ItemGroup>
     <!-- Sample XAML -->
-    <Page Include="Samples\Layers\DisplayKmlTours\DisplayKmlTours.xaml">
+    <Page Include="Samples\Layers\PlayKmlTours\PlayKmlTours.xaml">
       <Generator>MSBuild:Compile</Generator>
       <SubType>Designer</SubType>
     </Page>
@@ -1253,7 +1253,7 @@
   </ItemGroup>
   <ItemGroup>
     <!-- Screenshots -->
-    <Content Include="Samples\Layers\DisplayKmlTours\DisplayKmlTours.jpg">
+    <Content Include="Samples\Layers\PlayKmlTours\PlayKmlTours.jpg">
       <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
     </Content>
     <Content Include="Samples\Map\OpenMobileScenePackage\OpenMobileScenePackage.jpg">
