+++ src/WPF/ArcGISRuntime.WPF.Viewer/Samples/Layers/PlayKmlTours/PlayKmlTours.xaml
@@ -1,4 +1,4 @@
-﻿<UserControl x:Class="ArcGISRuntime.WPF.Samples.DisplayKmlTours.DisplayKmlTours"
+﻿<UserControl x:Class="ArcGISRuntime.WPF.Samples.PlayKmlTours.DisPlayKmlTours
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              xmlns:esri="http://schemas.esri.com/arcgis/runtime/2013">
