+++ src/WPF/ArcGISRuntime.WPF.Viewer/Samples/Layers/PlayKmlTours/PlayKmlTours.xaml.cs
@@ -3,34 +3,33 @@
 // Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with the License.
 // You may obtain a copy of the License at: http://www.apache.org/licenses/LICENSE-2.0
 //
-// Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on an 
-// "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the specific 
+// Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on an
+// "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the specific
 // language governing permissions and limitations under the License.
 
-using ArcGISRuntime.Samples.Managers;
 using Esri.ArcGISRuntime.Mapping;
 using Esri.ArcGISRuntime.Ogc;
 using System;
 using System.Collections.Generic;
 using System.Diagnostics;
 using System.Linq;
-using Windows.UI.Popups;
-using Windows.UI.Xaml;
+using System.Windows;
+using ArcGISRuntime.Samples.Managers;
 
-namespace ArcGISRuntime.UWP.Samples.DisplayKmlTours
+namespace ArcGISRuntime.WPF.Samples.PlayKmlTours
 {
     [ArcGISRuntime.Samples.Shared.Attributes.Sample(
-        "Display KML tours",
+        "Play a KML tour",
         "Layers",
-        "Display tours in KML files.",
+        "Play tours in KML files.",
         "")]
     [ArcGISRuntime.Samples.Shared.Attributes.OfflineData("f10b1d37fdd645c9bc9b189fb546307c")]
-    public partial class DisplayKmlTours
+    public partial class PlayKmlTours
     {
         // The KML tour controller provides player controls for KML tours.
         private readonly KmlTourController _tourController = new KmlTourController();
 
-        public DisplayKmlTours()
+        public PlayKmlTours()
         {
             InitializeComponent();
             Initialize();
@@ -76,7 +75,7 @@ namespace ArcGISRuntime.UWP.Samples.DisplayKmlTours
             catch (Exception e)
             {
                 Debug.WriteLine(e.ToString());
-                await new MessageDialog(e.ToString(), "Error").ShowAsync();
+                MessageBox.Show(e.ToString(), "Error");
             }
         }
 
@@ -118,7 +117,7 @@ namespace ArcGISRuntime.UWP.Samples.DisplayKmlTours
             }
         }
 
-        private void Play_Clicked(object sender, RoutedEventArgs routedEventArgs)
+        private void Play_Clicked(object sender, EventArgs e)
         {
             if (_tourController != null)
             {
@@ -132,7 +131,7 @@ namespace ArcGISRuntime.UWP.Samples.DisplayKmlTours
             }
         }
 
-        private void Pause_Clicked(object sender, RoutedEventArgs routedEventArgs)
+        private void Pause_Clicked(object sender, EventArgs e)
         {
             if (_tourController != null)
             {
@@ -145,7 +144,7 @@ namespace ArcGISRuntime.UWP.Samples.DisplayKmlTours
             }
         }
 
-        private void Reset_Clicked(object sender, RoutedEventArgs routedEventArgs)
+        private void Reset_Clicked(object sender, EventArgs e)
         {
             if (_tourController != null)
             {
