+++ src/iOS/Xamarin.iOS/ArcGISRuntime.Xamarin.Samples.iOS.csproj
@@ -236,7 +236,7 @@
   <ItemGroup>
     <!-- Screenshots -->
 	  <None Include="Samples\Map\OpenMobileScenePackage\OpenMobileScenePackage.jpg" />
-    <None Include="Samples\Layers\DisplayKmlTours\DisplayKmlTours.jpg" />
+    <None Include="Samples\Layers\PlayKmlTours\PlayKmlTours.jpg" />
     <None Include="Samples\Map\MapReferenceScale\MapReferenceScale.jpg" />
 	  <None Include="Samples\Map\CreateTerrainSurfaceFromRaster\CreateTerrainSurfaceFromRaster.jpg" />
 	  <None Include="Samples\Map\CreateTerrainSurfaceFromTilePackage\CreateTerrainSurfaceFromTilePackage.jpg" />
@@ -389,7 +389,7 @@
   <ItemGroup>
     <!-- Sample Code -->
 	  <Compile Include="Samples\Map\OpenMobileScenePackage\OpenMobileScenePackage.cs" />
-    <Compile Include="Samples\Layers\DisplayKmlTours\DisplayKmlTours.cs" />
+    <Compile Include="Samples\Layers\PlayKmlTours\PlayKmlTours.cs" />
     <Compile Include="Samples\Map\MapReferenceScale\MapReferenceScale.cs" />
 	  <Compile Include="Samples\Map\CreateTerrainSurfaceFromRaster\CreateTerrainSurfaceFromRaster.cs" />
 	  <Compile Include="Samples\Map\CreateTerrainSurfaceFromTilePackage\CreateTerrainSurfaceFromTilePackage.cs" />
