+++ src/iOS/Xamarin.iOS/Samples/Layers/PlayKmlTours/PlayKmlTours.jpg
